fixed project name after merge
