--- src/main/java/com/mycompany/app/config/CorsConfig.java
@@ -1,2 +1,0 @@
-package com.mycompany.app.config;public class CorsConfig {
-}
