--- src/main/java/com/mycompany/app/validation/NoHtml.java
@@ -1,15 +1,0 @@
-package com.mycompany.app.validation;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-import java.lang.annotation.*;
-
-@Documented
-@Constraint(validatedBy = NoHtmlValidator.class)
-@Target({ElementType.METHOD, ElementType.FIELD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface NoHtml {
-    String message() default "HTML tags are not allowed";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
