+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/config/CorsConfig.java
@@ -1,0 +1,2 @@
+package org.orange.oie.internship2025.assetmanagementsystem.config;public class CorsConfig {
+}
