+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/controller/AuthController.java
@@ -1,0 +1,24 @@
+package org.orange.oie.internship2025.assetmanagementsystem.controller;
+
+import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDTO;
+import org.orange.oie.internship2025.assetmanagementsystem.serivce.AuthService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/auth")
+public class AuthController {
+
+    private final AuthService authService;
+
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<UserDTO> authenticateUser(Authentication authentication) {
+        System.out.println("hello");
+        return authService.authenticateUser(authentication);
+    }
+}
