+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/validation/NoHtml.java
@@ -1,0 +1,15 @@
+package org.orange.oie.internship2025.assetmanagementsystem.validation;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+import java.lang.annotation.*;
+
+@Documented
+@Constraint(validatedBy = NoHtmlValidator.class)
+@Target({ElementType.METHOD, ElementType.FIELD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface NoHtml {
+    String message() default "HTML tags are not allowed";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}
